Fix crash when showing group without dedicated photo
